+++ lib/cache/get-stat.js
@@ -1,12 +1,10 @@
 'use strict';
-/* jshint node: true */
 
-var fs = require("fs")
-
-var npm = require("../npm.js")
-var log = require("npmlog")
 var mkdir = require("mkdirp")
-var chownr = require("chownr")
+  , fs = require("graceful-fs")
+  , log = require("npmlog")
+  , chownr = require("chownr")
+  , npm = require("../npm.js")
 
 function makeCacheDir (cb) {
   if (!process.getuid) return mkdir(npm.cache, cb)
