+++ lib/utils/depr-check.js
@@ -1,0 +1,15 @@
+'use strict';
+
+var log = require("npmlog")
+
+var deprecated = {}
+  , deprWarned = {}
+module.exports = function deprCheck (data) {
+  if (deprecated[data._id]) data.deprecated = deprecated[data._id]
+  if (data.deprecated) deprecated[data._id] = data.deprecated
+  else return
+  if (!deprWarned[data._id]) {
+    deprWarned[data._id] = true
+    log.warn("deprecated", "%s: %s", data._id, data.deprecated)
+  }
+}
